requirements.txt
pip
pyserial